Combine AppRepository and AppDataManager
